# build: speed up locale file generation to improve DX