--- a/packages/common/locales/generate-locales-tool/bin/write-locale-files-to-dist.ts
+++ b/packages/common/locales/generate-locales-tool/bin/write-locale-files-to-dist.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {writeFileSync} from 'fs';
+import fs from 'fs';
 import {join} from 'path';
 
 import {CldrData} from '../cldr-data';
@@ -21,7 +21,7 @@ import {BASE_LOCALE} from './base-locale';
  * Generates locale files for each available CLDR locale and writes it to the
  * specified directory.
  */
-function main(outputDir: string|undefined) {
+async function main(outputDir: string|undefined) {
   if (outputDir === undefined) {
     throw Error('No output directory specified.');
   }
@@ -35,16 +35,21 @@ function main(outputDir: string|undefined) {
   console.info(`Writing locales to: ${outputDir}`);
 
   // Generate locale files for all locales we have data for.
-  cldrData.availableLocales.forEach((localeData) => {
+  await Promise.all(cldrData.availableLocales.map(async (localeData) => {
     const locale = localeData.locale;
     const localeFile = generateLocale(locale, localeData, baseCurrencies);
     const localeExtraFile = generateLocaleExtra(locale, localeData);
     const localeGlobalFile = generateLocaleGlobalFile(locale, localeData, baseCurrencies);
 
-    writeFileSync(join(outputDir, `${locale}.ts`), localeFile);
-    writeFileSync(join(extraLocaleDir, `${locale}.ts`), localeExtraFile);
-    writeFileSync(join(globalLocaleDir, `${locale}.js`), localeGlobalFile);
-  });
+    await Promise.all([
+      fs.promises.writeFile(join(outputDir, `${locale}.ts`), localeFile),
+      fs.promises.writeFile(join(extraLocaleDir, `${locale}.ts`), localeExtraFile),
+      fs.promises.writeFile(join(globalLocaleDir, `${locale}.js`), localeGlobalFile),
+    ]);
+  }));
 }
 
-main(process.argv[2]);
+main(process.argv[2]).catch(err => {
+  console.error(err);
+  process.exitCode = 1;
+})