--- a/packages/common/locales/generate-locales-tool/plural-function.ts
+++ b/packages/common/locales/generate-locales-tool/plural-function.ts
@@ -12,6 +12,9 @@ import {CldrLocaleData} from './cldr-data';
 // There are no types available for `cldr`.
 const {load: createCldr} = await import('cldr' as any);
 
+// Load once to avoid re-parsing CLDR XML data on every invocation.
+const cldr = createCldr(runfiles.resolve('cldr_xml_data'));
+
 /**
  * Returns the plural function for a locale.
  */
@@ -21,7 +24,6 @@ export function getPluralFunction(localeData: CldrLocaleData, withTypes = true) 
   // we follow the CLDR-specified bundle lookup algorithm. A language does not necessarily
   // resolve directly to a bundle CLDR provides data for.
   const bundleName = localeData.attributes.bundle;
-  const cldr = createCldr(runfiles.resolve('cldr_xml_data'));
   let fn = cldr.extractPluralRuleFunction(bundleName).toString();
 
   const numberType = withTypes ? ': number' : '';