Fixes the bottleneck of locale generation impacting DX negatively, after auditing the common workflows

Before this commit, building everything to run `@angular/core` tests (`bazel clean` first) took:

```
INFO: Elapsed time: 76.496s, Critical Path: 72.92s
INFO: 225 processes: 125 internal, 5 linux-sandbox, 2 local, 93 worker.
INFO: Build completed successfully, 225 total actions
```

After:
```
INFO: Elapsed time: 41.756s, Critical Path: 38.18s
INFO: 225 processes: 125 internal, 5 linux-sandbox, 2 local, 93 worker.
INFO: Build completed successfully, 225 total actions
```

**76.5s to 41s. Clean Bazel build. 46.4% time reduction!**

Locale generation itself reduction: 69 -> 17s: 75.3% reduction

This being on a specialist Cloudtop.